requirements.txt
streamlit==1.39.0
streamlit-tags==1.2.8
python-dotenv==1.0.1
openai==1.53.0
token-count==0.2.1
beautifulsoup4==4.12.3
selenium==4.26.1
webdriver-manager==4.0.2
html2text==2024.2.26
pymongo==4.10.1
streamlit==1.39.0
streamlit-tags==1.2.8
streamlit-option-menu==0.4.0
openai==1.53.0